[ValueObjects] moved value objects to a single file
